backend/requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==3.7.1
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
fast-flights==2.2
fastapi==0.104.1
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.25.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kagglehub==1.0.0
kagglesdk==0.1.16
MarkupSafe==3.0.3
narwhals==2.18.1
numpy==2.4.3
packaging==26.0
pillow==12.1.1
playwright==1.58.0
primp==1.1.3
protobuf==6.33.6
pyarrow==23.0.1
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.9.1
pyee==13.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2026.1.post1
PyYAML==6.0.3
rebrowser_playwright==1.52.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
selectolax==0.4.7
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
starlette==0.27.0
streamlit==1.55.0
streamlit-searchbox==0.1.24
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.24.0
watchdog==6.0.0
websockets==16.0
